+++ app/include/view/player_setting.hpp
@@ -21,6 +21,14 @@ public:
     inline static int selectedSubtitle = 0;
     inline static int selectedAudio = 0;
 
+    enum class Equalizer {
+        BRIGHTNESS,
+        CONTRAST,
+        SATURATION,
+        HUE,
+        GAMMA,
+    };
+
 private:
     BRLS_BIND(brls::ScrollingFrame, settings, "player/settings");
     BRLS_BIND(brls::Box, cancel, "player/cancel");
@@ -35,4 +43,15 @@ private:
     BRLS_BIND(brls::SelectorCell, btnVideoMirror, "setting/video/mirror");
     BRLS_BIND(brls::SelectorCell, btnVideoAspect, "setting/video/aspect");
     BRLS_BIND(brls::SliderCell, btnSubsync, "setting/video/subsync");
+    // equalizer setting
+    BRLS_BIND(brls::RadioCell, btnEqualizerReset, "setting/equalizer/reset");
+    BRLS_BIND(brls::SliderCell, btnEqualizerBrightness, "setting/equalizer/brightness");
+    BRLS_BIND(brls::SliderCell, btnEqualizerContrast, "setting/equalizer/contrast");
+    BRLS_BIND(brls::SliderCell, btnEqualizerSaturation, "setting/equalizer/saturation");
+    BRLS_BIND(brls::SliderCell, btnEqualizerGamma, "setting/equalizer/gamma");
+    BRLS_BIND(brls::SliderCell, btnEqualizerHue, "setting/equalizer/hue");
+
+    void setupEqualizer(brls::SliderCell* cell, const std::string& title, Equalizer item, double initValue);
+
+    void registerHideBackground(brls::View* view);
 };